feat(gui): add unified batch processing GUI with job queue management
- Replace previous gui_batch.py with unified src/gui.py for batch STL processing
- Integrate EnhancedJobManager for batch job queue handling
- Add drag-and-drop and browse buttons for adding STL files/folders to queue
- Implement batch queue tab with job list, controls (start, pause, stop, clear)
- Add periodic UI updates reflecting job queue state and progress
- Provide menu options and keyboard shortcuts for batch operations
- Include batch processing help dialog and improved error handling
- Retain single file processing tabs (Analysis, Validation, Rendering, Generators)
- Update setup.py to point stl-gui console script to new unified GUI
- Update documentation references from gui_batch to gui

This commit introduces a comprehensive batch processing system into the GUI, enabling users to manage multiple STL processing jobs efficiently within a single unified interface.
